Fix compilation warnings in core sources
Agent-Logs-Url: https://github.com/arkq/android-matter-controller/sessions/d2698a9f-0aaa-4651-a60e-a830ce1d6f2a

+++ app/src/main/java/io/aether/android/chip/ClustersHelper.kt
@@ -89,7 +89,7 @@ class ClustersHelper @Inject constructor(private val chipClient: ChipClient) {
       Timber.d("part [$part] is [${part.javaClass}]")
       val endpointInt =
           when (part) {
-            is Int -> part.toInt()
+            is Int -> part
             else -> return@forEach
           }
       Timber.d("Processing part [$part]")
